Merge de ramas

+++ Código/3DM.hpp
@@ -4,7 +4,8 @@
 #include <stdlib.h>
 #include <cmath>
 #include "tripleta.hpp"
-#include <cassert>
+
+#include "bitVector.hpp"
 
 using namespace std;
 
@@ -24,7 +25,7 @@ class _3DM {
         vector<string> getZ() const;
         vector<tripleta> getW() const;
 
-        vector<int> transformToPartition();
+        vector<long int> transformToPartition();
 };
 
 ostream& operator << (ostream& os, const _3DM &p);
